feat: Add YAML to HTML converter in jqlang
This is inspired by <https://github.com/RickCole21/json-to-html>. The functionality is the same, except default keys are different, and options are read via --arg

The idea is the same, but to use YAML I have to use 'yq to jq' pipe. Not great. And I probably should make defaults the same as they used to be. IDK.'