feat: Add YAML templating
This is mostly a PoC. We use custom tags in YAML file to hint that we want this blocks preprocessed. JQlang has better syntax for complex operations, but YAML itself is quite flexible, so more funny stuff can be achieved. For example, YQ has 'eval', and we can expose it to the user. Fun! Happy debugging